# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank API
  description: API for managing bank cards
  version: 1.0.0
  contact:
    name: Timofey Zaytsev
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates user and returns access and refresh tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Returns refresh and access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Auth exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /auth/refresh:
    put:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generates new access token using refresh token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New access token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '400':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Auth exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
...
  /admin/users/{userId}:
    get:
      tags:
        - Admin Users
      summary: Get user by ID
      description: Retrieves user information by user ID (admin only)
      operationId: getUserById
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindUserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /admin/users:
    get:
      tags:
        - Admin Users
      summary: Get all users
      description: Retrieves paginated list of users with optional filtering
      operationId: findAllUsers
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (0-indexed)
          example: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
          description: Number of items per page
          example: 20
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search term for filtering users by email, full name. Supports partial matching (e.g., 'john' will match 'john.doe@example.com')
          example: "john"
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserRole'
          description: Filter by user role
          example: "ADMIN"
        - name: isEnabled
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by account status (true for enabled accounts, false for disabled)
          example: true
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/FindUserResponse'
                  pageable:
                    type: object
                    properties:
                      pageNumber:
                        type: integer
                        description: Current page number
                      pageSize:
                        type: integer
                        description: Number of items per page
                      sort:
                        type: object
                        description: Sort information
                      offset:
                        type: integer
                        description: Offset from the start
                      paged:
                        type: boolean
                        description: Whether pagination is enabled
                      unpaged:
                        type: boolean
                        description: Whether pagination is disabled
                  totalPages:
                    type: integer
                    description: Total number of pages
                  totalElements:
                    type: integer
                    description: Total number of elements
                  last:
                    type: boolean
                    description: Whether this is the last page
                  size:
                    type: integer
                    description: Number of items per page
                  number:
                    type: integer
                    description: Current page number
                  sort:
                    type: object
                    description: Sort information
                  first:
                    type: boolean
                    description: Whether this is the first page
                  numberOfElements:
                    type: integer
                    description: Number of elements in the current page
                  empty:
                    type: boolean
                    description: Whether the page is empty
        '400':
          description: Validation exception (invalid page or size parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
    post:
      tags:
        - Admin Users
      summary: Create user
      description: Creates a new user (admin only)
      operationId: createUser
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /admin/users/{userId}/block:
    put:
      tags:
        - Admin Users
      summary: Block user
      description: Blocks a user by ID (admin only)
      operationId: blockUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '204':
          description: User blocked successfully (enable = false)
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description:  email
          example: "john.doe@mail.com"
        password:
          type: string
          format: password
          description: User password
          example: "SecurePass123!"
          writeOnly: true
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: JWT refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expiresIn:
          type: integer
          format: long
          description: Access token expiration time in seconds
          example: 3600
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    AccessTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: New JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    FindUserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: User ID
          example: "3ab2c3d4-e5f6-7890-abcd-ef1234567890"
        fullName:
          type: string
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User email
          example: "john.doe@example.com"
        createdAt:
          type: string
          format: date-time
          description: User creation timestamp
          example: "2026-01-01T12:00:00Z"
        enabled:
          type: boolean
          description: Whether the user is enabled
          example: true
        roleNames:
          type: array
          description: List of user roles
          items:
            $ref: '#/components/schemas/UserRole'
    CreateUserRequest:
      type: object
      required:
        - email
        - fullName
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address. Must be unique and valid format.
          maxLength: 100
          example: "john.doe@example.com"
        fullName:
          type: string
          description: User's full name. Can include first and last name.
          minLength: 2
          maxLength: 100
          pattern: "^[a-zA-Zа-яА-Я\\s'-]+$"
          example: "John Doe"
        password:
          type: string
          format: password
          description: |
            User's password. Must meet security requirements: 
            at least 8 characters, contain at least one digit, 
            one lowercase, one uppercase, and one special character (@#$%^&+=!).
          minLength: 8
          maxLength: 100
          pattern: "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=!]).*$"
          example: "SecurePass123!"
          writeOnly: true
        isAdmin:
          type: boolean
          format: boolean
          description: Flag indicating whether the user should be created with administrator privileges. 
           If set to true, the user will have ADMIN role in addition to the default USER role. 
           Default value is false.
          example: false
...
    CreateUserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: Created user ID
          example: "123e4567-e89b-12d3-a456-426614174000"
...
    UserRole:
      type: string
      enum:
        - USER
        - ADMIN
        - MANAGER
      description: User role type
...
    ErrorResponseDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad Request"
        uri:
          type: string
          example: "/api/v1/admin/cards"